A lot of refactoring

+++ SPOStackMachine/src/main/java/ru/therealmone/SPOStackMachine/Element.java
@@ -3,20 +3,20 @@ package ru.therealmone.SPOStackMachine;
 import ru.therealmone.TranslatorAPI.Node;
 
 class Element extends Node {
-    private int value;
+    private double value;
     private int hashCode;
 
-    Element(String name, int value) {
+    Element(String name, double value) {
         super(name);
         this.value = value;
         this.hashCode = name.hashCode();
     }
 
-    int getValue() {
+    double getValue() {
         return this.value;
     }
 
-    void setValue(int value) {this.value = value;}
+    void setValue(double value) {this.value = value;}
 
     int getHashCode() {
         return this.hashCode;
